+++ server/src/routes/v1/__init__.py
@@ -1,4 +1,5 @@
 from fastapi import APIRouter
+from .lists import router as lists_router
 
 router = APIRouter()
 
@@ -6,3 +7,6 @@ router = APIRouter()
 @router.get("/")
 def index():
     return {"msg": "You have successfully reached API v1"}
+
+
+router.include_router(lists_router)
